Corrected issues in the gradients
There were some very subtle issues connected with computing the gradients. They popped out only in the cases where we go beyond the isomorphism between the data graph and the graph of each SPSN component.

When resolving these problems, simplified also the implmentation of the ProductNode and the associated _reflectinmodel.